+++ testing/trino-testing/src/main/java/io/trino/testing/MultisetAssertions.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.plugin.hive.util;
+package io.trino.testing;
 
 import com.google.common.collect.Multiset;
 import com.google.common.collect.Sets;
